==> capstone/code/1.2_processing_data.ipynb <==
## 1.2. Processing Data

**Split by weeks (part 2)**

This part is solely to split the training click/cart/order data by weeks as data is too huge for analysis or preprocessing. It was separate from Part 1 due to memory issues.

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import pyarrow.parquet as pq

import datetime

# import parquet files
train_click = pd.read_parquet('/content/drive/MyDrive/0.capstone/train_click.parquet')
train_cart = pd.read_parquet('/content/drive/MyDrive/0.capstone/train_cart.parquet')
train_order = pd.read_parquet('/content/drive/MyDrive/0.capstone/train_order.parquet')

# function to filter by week. used query instead of loc as it is more memory efficient
def filter_by_week(df, start_unixtime, end_unixtime):
    return df.query('@start_unixtime <= ts < @end_unixtime').reset_index(drop=True)

train_click_1 = filter_by_week(train_click, 1659304800, 1659909600)
train_click_1.to_parquet('/content/drive/MyDrive/0.capstone/train_click_1.parquet')

train_click_2 = filter_by_week(train_click, 1659909600, 1660514400)
train_click_2.to_parquet('/content/drive/MyDrive/0.capstone/train_click_2.parquet')

train_click_3 = filter_by_week(train_click, 1660514400, 1661119200)
train_click_3.to_parquet('/content/drive/MyDrive/0.capstone/train_click_3.parquet')

train_click_4 = filter_by_week(train_click, 1661119200, 1661724000)
train_click_4.to_parquet('/content/drive/MyDrive/0.capstone/train_click_4.parquet')

train_cart_1 = filter_by_week(train_cart, 1659304800, 1659909600)
train_cart_1.to_parquet('/content/drive/MyDrive/0.capstone/train_cart_1.parquet')

train_cart_2 = filter_by_week(train_cart, 1659909600, 1660514400)
train_cart_2.to_parquet('/content/drive/MyDrive/0.capstone/train_cart_2.parquet')

train_cart_3 = filter_by_week(train_cart, 1660514400, 1661119200)
train_cart_3.to_parquet('/content/drive/MyDrive/0.capstone/train_cart_3.parquet')

train_cart_4 = filter_by_week(train_cart, 1661119200, 1661724000)
train_cart_4.to_parquet('/content/drive/MyDrive/0.capstone/train_cart_4.parquet')

train_order_1 = filter_by_week(train_order, 1659304800, 1659909600)
train_order_1.to_parquet('/content/drive/MyDrive/0.capstone/train_order_1.parquet')

train_order_2 = filter_by_week(train_order, 1659909600, 1660514400)
train_order_2.to_parquet('/content/drive/MyDrive/0.capstone/train_order_2.parquet')

train_order_3 = filter_by_week(train_order, 1660514400, 1661119200)
train_order_3.to_parquet('/content/drive/MyDrive/0.capstone/train_order_3.parquet')

train_order_4 = filter_by_week(train_order, 1661119200, 1661724000)
train_order_4.to_parquet('/content/drive/MyDrive/0.capstone/train_order_4.parquet')

# print to check number of rows, that split is done correctly
print(train_click_1.shape[0] + train_click_2.shape[0] + train_click_3.shape[0] + train_click_4.shape[0])
print(train_cart_1.shape[0] + train_cart_2.shape[0] + train_cart_3.shape[0] + train_cart_4.shape[0])
print(train_order_1.shape[0] + train_order_2.shape[0] + train_order_3.shape[0] + train_order_4.shape[0])
