==> capstone/code/1.3_more_processing_data.ipynb <==
## 1.3. More Processing Data

**Split by further chunks**

This part is solely to split the training click/cart/order data by smaller chunks for processing later. It was separate from Part 1 due to memory issues.

import pandas as pd, numpy as np

from google.colab import drive
drive.mount('/content/drive')

train = pd.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/validation/train_parquet/train.parquet')
test = pd.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/validation/test_parquet/test.parquet')

train.head()

type_labels = {0: 'clicks', 1: 'carts', 2: 'orders'}
train['type'] = train['type'].replace(type_labels)

train.head()

# split train into chunk size of 100
chunk_size = len(train) // 100 + 1
for i, chunk in enumerate(range(0, len(train), chunk_size)):
    filename = f"{i+1:03}.parquet"  # create filename with leading zeros
    df_chunk = train.iloc[chunk:chunk+chunk_size]
    output_file = '/content/drive/MyDrive/0.capstone/covisit_prepare/validation/train_parquet/' + filename
    df_chunk.to_parquet(output_file, index=False)

type_labels = {0: 'clicks', 1: 'carts', 2: 'orders'}
test['type'] = test['type'].replace(type_labels)

test.head()

# split test into chunk size of 20
chunk_size = len(test) // 20 + 1
for i, chunk in enumerate(range(0, len(test), chunk_size)):
    filename = f"{i+1:03}.parquet"  # create filename with leading zeros
    df_chunk = test.iloc[chunk:chunk+chunk_size]
    output_file = '/content/drive/MyDrive/0.capstone/covisit_prepare/validation/test_parquet/' + filename
    df_chunk.to_parquet(output_file, index=False)
