==> capstone/code/5.0_validate_intuition.ipynb <==
## 5.0. Validate intuition

Running the full code is time and resource intensive. Thus, we designed a quick way to validate some ideas. There will be huge overfitting, but it works fine for testing some ideas.

Ideas tested:
1. Open carts. Intuitively, open carts are most likely to lead to orders.
2. Repeated orders. Intuitively, repeated orders of the item by user should mean that they need to buy that item often.
3. Recent carts/orders, carts and orders.

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np

from collections import Counter

train_cart_4 = pd.read_parquet('/content/drive/MyDrive/0.capstone/train_cart_4.parquet')

train_order_wo_val_4 = pd.read_parquet('/content/drive/MyDrive/0.capstone/train_val/train_order_wo_val_4.parquet')
val_order_4 = pd.read_parquet('/content/drive/MyDrive/0.capstone/train_val/val_order_4.parquet')

**Create functions**

# Function to provide validation score
def validate_recall(validation):
  # Calculate the recall for each session
  recalls = []
  for session, group in validation.groupby('session'):
      ground_truth = Counter(group['aid_x'].explode())
      predicted = Counter(group['aid_y'].explode())
      intersection_count = sum((ground_truth & predicted).values())
      ground_truth_count = sum(ground_truth.values())
      if ground_truth_count > 0:
          recalls.append(intersection_count / ground_truth_count)

  # Aggregate mean of recall
  aggregate_recall = sum(recalls) / len(recalls)
  print(f'Validation recall score: {aggregate_recall}')

def merge_two(df_left, df_right):
  df_new = df_left.merge(df_right, on='session', how='left')
  return df_new[df_new['aid_y'].notna()]

**1. Validate list of open carts**

cart_order_4 = train_cart_4.merge(train_order_wo_val_4, how='left', on=['session', 'aid'])

# open carts where there are no orders or orders timestamp is not later than carts timestamp
cart_order_4 = cart_order_4[(pd.isna(cart_order_4['ts_y'])) | (cart_order_4['ts_x'] > cart_order_4['ts_y'])]

# keep only session and aid columns; drop duplicates
cart_order_4 = cart_order_4.loc[:, ['session', 'aid']].drop_duplicates(subset=['session', 'aid'], keep='first')

# create list of items by session
cart_order_4 = cart_order_4.groupby('session')['aid'].apply(list).reset_index(name='aid')

# Merge by session. x = ground truth ; y = predict
validation_open_carts = merge_two(val_order_4, cart_order_4)
validation_open_carts.head()

validate_recall(validation_open_carts)

Good validation score from only open orders

**2. Validate list of repeated orders**

# Group dataframe by session and create a list of all aids for each session
sessions = train_order_wo_val_4.groupby('session')['aid'].apply(list)

# Create counter for each aid in each session
aid_counts = {}
for session, aids in sessions.items():
    aid_counts[session] = Counter(aids)

# Create a dictionary of repeated aids for each session
repeated_aids = {}
for session, counts in aid_counts.items():
    repeated = {aid: count for aid, count in counts.items() if count > 1}
    if len(repeated) > 0:
        repeated_aids[session] = repeated

# Convert to dataframe
repeated_aids = {k: list(set(v)) for k, v in repeated_aids.items()}
repeated_aids_df = pd.DataFrame(repeated_aids.items(), columns=['session', 'aid'])
repeated_aids_df.head()

# Merge by session. x = ground truth ; y = predict
validation_repeated_orders = merge_two(val_order_4, repeated_aids_df)
validation_repeated_orders.head()

validate_recall(validation_repeated_orders)

Only a small number of repeated orders and the contribution to score is small.

**3.1. Recent 20 cart/orders**

# Combine data
recent = pd.concat([train_order_wo_val_4, train_cart_4], ignore_index=True)
recent = recent.sort_values(['session','ts'])
recent.head()

# Keep only last 20 unique aids per session
recent = recent.groupby('session').aid.unique().apply(lambda x: x[-20:]).reset_index()
recent.head()

# Merge by session. x = ground truth ; y = predict
validation_recent_buys = merge_two(val_order_4, recent)
validation_recent_buys.head()

validate_recall(validation_recent_buys)

Gives good score

**3.2. Recent 20 orders only**

recent_2 = train_order_wo_val_4
recent_2 = recent_2.sort_values(['session','ts'])
recent_2 = recent_2.groupby('session').aid.unique().apply(lambda x: x[-20:]).reset_index()
recent_2.head()

# Merge by session. x = ground truth ; y = predict
validation_recent_orders = merge_two(val_order_4, recent_2)
validation_recent_orders.head()

validate_recall(validation_recent_orders)

Recent orders did not really give good score.

3.3. Recent 20 carts only

recent_3 = train_cart_4
recent_3 = recent_3.sort_values(['session','ts'])
recent_3 = recent_3.groupby('session').aid.unique().apply(lambda x: x[-20:]).reset_index()
recent_3.head()

# Merge by session. x = ground truth ; y = predict
validation_recent_carts = merge_two(val_order_4, recent_3)
validation_recent_carts.head()

validate_recall(validation_recent_carts)

Did slightly better than carts and orders

**Summary**
- Recent carts have high influence on next item. This is aligned with ranker models we have done.
- Open carts have high influence on next item. However, Kaggle score later showed that it didn't improve the model score. This could be because these are already captured under recent carts. Thus we did not include this in the model.
- Repeated orders show to have small influence on prediction. This is later seen as well in the boosting reranker models, where the feature importance is small.
