==> capstone/code/6.3_reranker_add_features.ipynb <==
## 6.3. Add features to candidates for Reranker

We need to add features to train and test candidates to fit and predict on the ranker. The 15 features we selected are as follows:

**Key user-item feature:**
- rank according to candidate generation sequence ("buy rank")

This is to rank according to earlier rankings in Parts 6.1 / 6.2. Idea is to use this as a simple way to account for recency, type, frequency and covistation like how the rule based model does. Ideally, the model will weigh this feature as heavily important.

**User features:**
- total number of interactions for each user
- number of unique items for each user
- user session length

Intuition is that users who have a lot of activities are likely to make a next purchase.

**Item features:**
- number of unique users for each item
- total interactions for each item (carts only and orders only)
- number of co-visitation pairs

Intuition is that these items are more likely to be a popular items, thus higher chances of being within next 20 orders 

**Item feature:**
- cart_order_interaction_ratio

Intuition is that if there were high cart-order conversion, the item will have higher chances of being within the next 20 orders.



**User-item features:**
- recurring items (incl carts only and orders only): intuition is that items that were bought often by users are more likely to be purchased again, e.g. out of necessity
- last click / last cart: intuition is that the last click and last cart is most likely to have far more influence on the next action
<br><br>

---
<br>

**Use of Polars** <br>
We explore using Polars here. Polars is a DataFrame API built on Rust, a low-level systems programming language known for its performance and memory safety. Polars runs faster than pandas. However because it is newer, it has less documentation for finding how to do more obscure dataframe manipulation compared to Pandas.

from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd

# !pip install polars
import polars as pl

import seaborn as sns

candidates_train = pd.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/ranker_candidates_train_3.parquet')
candidates_test = pd.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/ranker_candidates_test_3.parquet')

validation_data_a = pl.read_parquet('/content/drive/MyDrive/0.capstone/train_val/test.parquet')

train = pl.read_parquet('/content/drive/MyDrive/0.capstone/train.parquet')
test = pl.read_parquet('/content/drive/MyDrive/0.capstone/test.parquet')

top_30_buy2buy_train = pl.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/top_30_buy2buy_v4_0.parquet')
top_30_buy_train_1 = pl.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/top_30_carts_orders_v4_0.parquet')
top_30_buy_train_2 = pl.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/top_30_carts_orders_v4_1.parquet')

top_30_buy2buy_test = pd.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/top_30_buy2buy_v3_0.parquet')
top_30_buy_test_1 = pd.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/top_30_carts_orders_v3_0.parquet')
top_30_buy_test_2 = pd.read_parquet('/content/drive/MyDrive/0.capstone/covisit_prepare/top_30_carts_orders_v3_1.parquet')

**Preprocessing**

# Function to align datatypes for merge
def change_type(df):
  df = df.astype({'aid_x': 'int32', 'aid_y': 'int32'})
  return pl.from_pandas(df)

# Align datatypes for merge
top_30_buy2buy_test = change_type(top_30_buy2buy_test)
top_30_buy_test_1 = change_type(top_30_buy_test_1)
top_30_buy_test_2 = change_type(top_30_buy_test_2)

# Concat train with validation_data_a for aid features
train_new = pl.concat([train, validation_data_a])

# Concat train with validation_data_a for aid features
test_new = pl.concat([train, test])

**1 Key user-item feature:** Rank according to candidate generation sequence

# Assign rank based on candidate generation sequence
# Convert to i32 dtype and then convert to polar
def buy_rank(df):
    df['buy_rank'] = 1 / (df.groupby('session').cumcount() + 1)
    df = df.astype({'session': 'int32', 'aid': 'int32', 'buy_rank': 'float64'})
    return pl.from_pandas(df)

candidates_train = buy_rank(candidates_train)
candidates_test = buy_rank(candidates_test)

print(candidates_train.dtypes)
print(candidates_test.dtypes)

print(candidates_train.shape)
print(candidates_test.shape)

candidates_train.head(35)

candidates_test.head()

**3 user features:** total interactions, number of unique aids, session length

test.head()

# Session features
def session_features(df):
  return df.select([
          pl.col('session'),
          pl.col('aid'),
          (pl.col('ts')*1000 + 2*60*60*1000*1000).cast(pl.Datetime).alias('timestamp'),
      ]).groupby('session').agg([
          pl.col('timestamp').count().cast(pl.Int16).alias('session_total_interactions'),
          pl.col('aid').n_unique().cast(pl.Int16).alias('session_unique_aids'),
          pl.col('timestamp').min().alias('transaction_start_timestamp'),
          pl.col('timestamp').max().alias('transaction_end_timestamp')
      ]).select([
          '*',
          ((pl.col('transaction_end_timestamp') - pl.col('transaction_start_timestamp')).apply(lambda x: x.total_seconds())).cast(pl.Float32).alias('session_length')
      ]).drop(['transaction_end_timestamp', 'transaction_start_timestamp'])

train_session_features = session_features(validation_data_a)

train_session_features.head()

test_session_features = session_features(test)

test_session_features.head()

**2 x 2 item features:** number of unique sessions and total interactions (carts only and orders only)

# Aid features
def aid_features(df):
  return df.select([
      pl.col('session'),
      pl.col('aid'),    
      (pl.col('ts')*1000 + 2*60*60*1000*1000).cast(pl.Datetime).cast(pl.Date).alias('timestamp'),
      pl.col('session').n_unique().over(pl.lit(1)).alias('total_sessions')
      ]).groupby('aid').agg([
          pl.col('session').n_unique().cast(pl.Int32).alias('aid_unique_sessions'),
          pl.col('timestamp').count().cast(pl.Int32).alias('aid_total_interactions'),
      ])

# Filter orders only
train_orders = train_new.filter(pl.col('type') == pl.lit(2))
test_orders = test_new.filter(pl.col('type') == pl.lit(2))

train_orders_aid_features = aid_features(train_orders)
train_orders_aid_features = train_orders_aid_features.rename({
  'aid_unique_sessions': 'order_aid_unique_sessions',
  'aid_total_interactions': 'order_aid_total_interactions'
})
train_orders_aid_features.head()

test_orders_aid_features = aid_features(test_orders)
test_orders_aid_features = test_orders_aid_features.rename({
  'aid_unique_sessions': 'order_aid_unique_sessions',
  'aid_total_interactions': 'order_aid_total_interactions'
})
test_orders_aid_features.head()

# Filter carts only
train_carts = train_new.filter(pl.col('type') == pl.lit(1))
test_carts = test_new.filter(pl.col('type') == pl.lit(1))

train_carts_aid_features = aid_features(train_carts)
train_carts_aid_features = train_carts_aid_features.rename({
  'aid_unique_sessions': 'cart_aid_unique_sessions',
  'aid_total_interactions': 'cart_aid_total_interactions'
})
train_carts_aid_features.head()

test_carts_aid_features = aid_features(test_carts)
test_carts_aid_features = test_carts_aid_features.rename({
  'aid_unique_sessions': 'cart_aid_unique_sessions',
  'aid_total_interactions': 'cart_aid_total_interactions'
})
test_carts_aid_features.head()

**2 item features:** number of co-visitation pairs 

top_30_buy2buy_train.head()

def covisit_pair(df):
  return df.groupby('aid_x').agg(pl.col('aid_y').n_unique()).rename({'aid_x': 'aid', 'aid_y': 'pair_count'})

train_covisit_1_feature = covisit_pair(top_30_buy2buy_train)
train_covisit_1_feature.head()

test_covisit_1_feature = covisit_pair(top_30_buy2buy_test)
test_covisit_1_feature.head()

# Concat parts
top_30_buy_train = pl.concat([top_30_buy_train_1, top_30_buy_train_2])
top_30_buy_test = pl.concat([top_30_buy_test_1, top_30_buy_test_2])

train_covisit_2_feature = covisit_pair(top_30_buy_train)
train_covisit_2_feature = train_covisit_2_feature.rename({'pair_count': 'pair_count_2'})
train_covisit_2_feature.head()

test_covisit_2_feature = covisit_pair(top_30_buy_test)
test_covisit_2_feature = test_covisit_2_feature.rename({'pair_count': 'pair_count_2'})
test_covisit_2_feature.head()

**1 x 3 session-aid feature:** recurring aids (incl carts only and orders only)

def session_aid_features(df):
  return df.select([
          pl.col('session'),
          pl.col('aid'),
      ]).groupby('session', 'aid').agg([
          (pl.count('aid').alias('num_recurring_aids')).cast(pl.Int16)
      ])

# Filter orders only
train_session_aid_orders = validation_data_a.filter(pl.col('type') == pl.lit(2))
test_session_aid_orders = test.filter(pl.col('type') == pl.lit(2))

train_session_aid_orders_features = session_aid_features(train_session_aid_orders)
train_session_aid_orders_features = train_session_aid_orders_features.rename({'num_recurring_aids': 'order_num_recurring_aids'})

test_session_aid_orders_features = session_aid_features(test_session_aid_orders)
test_session_aid_orders_features = test_session_aid_orders_features.rename({'num_recurring_aids': 'order_num_recurring_aids'})

# Filter carts only
train_session_aid_carts = validation_data_a.filter(pl.col('type') == pl.lit(2))
test_session_aid_carts = test.filter(pl.col('type') == pl.lit(2))

train_session_aid_carts_features = session_aid_features(train_session_aid_carts)
train_session_aid_carts_features = train_session_aid_carts_features.rename({'num_recurring_aids': 'cart_num_recurring_aids'})

test_session_aid_carts_features = session_aid_features(test_session_aid_carts)
test_session_aid_carts_features = test_session_aid_carts_features.rename({'num_recurring_aids': 'cart_num_recurring_aids'})

**1 x2 session-aid feature:** last click / cart

# filter only last cart by timestamp
train_last_cart = train_carts.sort('ts').groupby('session').tail(1)
test_last_cart = test_carts.sort('ts').groupby('session').tail(1)

# keep only relevant columns
train_last_cart = train_last_cart.select(['session', 'aid', 'type']).rename({'type': 'last_cart'})
test_last_cart = test_last_cart.select(['session', 'aid', 'type']).rename({'type': 'last_cart'})

train_last_cart.head()

# Filter clicks only
train_clicks = train_new.filter(pl.col('type') == pl.lit(0))
test_clicks = test_new.filter(pl.col('type') == pl.lit(0))

# filter only last click by timestamp
train_last_click = train_clicks.sort('ts').groupby('session').tail(1)
test_last_click = test_clicks.sort('ts').groupby('session').tail(1)

# keep only relevant columns
train_last_click = train_last_click.select(['session', 'aid', 'type']).rename({'type': 'last_click'})
test_last_click = test_last_click.select(['session', 'aid', 'type']).rename({'type': 'last_click'})

train_last_click = train_last_click.with_columns(pl.col('last_click').apply(lambda x: 1 if x == 0 else x))
test_last_click = test_last_click.with_columns(pl.col('last_click').apply(lambda x: 1 if x == 0 else x))

train_last_click.head()

**Join candidates with features**

train_candidates_features = (
    candidates_train
    .join(train_session_features, how='left', on='session')
    .join(train_orders_aid_features, how='left', on='aid')
    .join(train_carts_aid_features, how='left', on='aid')
    .join(train_covisit_1_feature, how='left', on='aid')
    .join(train_covisit_2_feature, how='left', on='aid')
    .join(train_session_aid_orders_features, how='left', on=['session', 'aid'])
    .join(train_session_aid_carts_features, how='left', on=['session', 'aid'])
    .join(train_last_cart, how='left', on=['session', 'aid'])
    .join(train_last_click, how='left', on=['session', 'aid'])
)

# create new column based on cart-order interaction ratio
train_candidates_features = train_candidates_features.select(
    "*",
    (pl.col("cart_aid_total_interactions") / pl.col("order_aid_total_interactions")).alias("cart_order_interaction_ratio")
)

train_candidates_features.head(50)

test_candidates_features = (
    candidates_test
    .join(test_session_features, how='left', on='session')
    .join(test_orders_aid_features, how='left', on='aid')
    .join(test_carts_aid_features, how='left', on='aid')
    .join(test_covisit_1_feature, how='left', on='aid')
    .join(test_covisit_2_feature, how='left', on='aid')
    .join(test_session_aid_orders_features, how='left', on=['session', 'aid'])
    .join(test_session_aid_carts_features, how='left', on=['session', 'aid'])
    .join(test_last_cart, how='left', on=['session', 'aid'])
    .join(test_last_click, how='left', on=['session', 'aid'])
)

# create new column based on cart-order interaction ratio
test_candidates_features = test_candidates_features.select(
    "*",
    (pl.col("cart_aid_total_interactions") / pl.col("order_aid_total_interactions")).alias("cart_order_interaction_ratio")
)

test_candidates_features.head(50)

print(train_candidates_features.shape)
print(test_candidates_features.shape)

# train_candidates_features.write_parquet("/content/drive/MyDrive/0.capstone/covisit_prepare/reranker_candidates_features_train_5.parquet", compression="snappy")
# test_candidates_features.write_parquet("/content/drive/MyDrive/0.capstone/covisit_prepare/reranker_candidates_features_test_5.parquet", compression="snappy")
